Turning on -Werror, and fixing all new errors produced by that. Attempting to suppress errors from 3rd party libraries by including them with -isystem, but having to add -Wno-strict-aliasing to work around a boost error that still leaks through.

+++ src/Gaffer/Context.cpp
@@ -179,7 +179,7 @@ std::string Context::substitute( const std::string &s ) const
 				padding++;
 				i++;
 			}
-			int frame = round( getFrame() );
+			int frame = (int)round( getFrame() );
 			std::ostringstream padder;
 			padder << std::setw( padding ) << std::setfill( '0' ) << frame;
 			result += padder.str();
